hw6/hasker/requirements.txt
Django==1.11.4
Pillow==4.2.1
psycopg2==2.7.3
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0